requirements.txt
numpy>=1.23
pandas>=1.5
matplotlib>=3.6
statsmodels>=0.13
requests>=2.28
beautifulsoup4>=4.12
python-dateutil>=2.8.2
